long recipe EUt, sub-tick oc, remove negative EUt

+++ src/main/java/gregtech/integration/hwyla/provider/RecipeLogicDataProvider.java
@@ -47,7 +47,7 @@ public class RecipeLogicDataProvider extends CapabilityDataProvider<AbstractReci
         NBTTagCompound subTag = new NBTTagCompound();
         subTag.setBoolean("Working", capability.isWorking());
         if (capability.isWorking()) {
-            subTag.setInteger("RecipeEUt", capability.getInfoProviderEUt());
+            subTag.setLong("RecipeEUt", capability.getInfoProviderEUt());
         }
         tag.setTag("gregtech.AbstractRecipeLogic", subTag);
         return tag;
