+++ src/test/java/gregtech/api/capability/impl/AbstractRecipeLogicTest.java
@@ -73,7 +73,7 @@ public class AbstractRecipeLogicTest {
             }
 
             @Override
-            protected boolean drawEnergy(int recipeEUt, boolean simulate) {
+            protected boolean drawEnergy(long recipeEUt, boolean simulate) {
                 return true;
             }
